Add functioniality to fetch 'today' from JS and use for new Training
This is unbelievably hard to achieve. And so far very indirect and
stateful: All I want is just the current date as a string and store it
with a new Training, but for that do I really need a new store field,
a two-step process to fetch it via ports (since the builtin time/date
handling es even more complicated) and then an extra handlers to trigger
that ritual on startup? Jesus...